feat: moved aurora components in extra file
